requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
matplotlib==3.7.2
